Component: StyleOpenStreet(create constance file for longitude and latitude)
