+++ frontend/src/pages/streetMap/StreetMap.page.tsx
@@ -11,17 +11,7 @@ import {
   Tooltip,
 } from "react-leaflet";
 import { LocationMarker } from "../../scripts/check_position";
-
-const bamenda: any = [5.914395, 10.129316];
-const bafoussan: any = [5.468774, 10.420834];
-const bertoua: any = [4.558081, 13.662206];
-const buea: any = [4.155587, 9.232463];
-const douala: any = [4.03222, 9.706715];
-const ebolowa: any = [2.891746, 11.15648];
-const garoua: any = [9.278875, 13.394429];
-const maroua: any = [10.588261, 14.350791];
-const nagoundere: any = [7.349664, 13.577051];
-const yaounde: any = [3.826985, 11.495974];
+import { LATLONG } from "../../scripts/lat_long";
 
 const SetViewOnClick = (animateRef: any) => {
   const map = useMapEvent("click", (e) => {
@@ -40,7 +30,7 @@ export const StreetMap = () => {
     <MapContainer
       className="z-0"
       style={{ height: "100vh" }}
-      center={yaounde}
+      center={LATLONG.yaounde}
       zoom={7}
       zoomControl={false}
       scrollWheelZoom={true}
@@ -49,45 +39,77 @@ export const StreetMap = () => {
         attribution='&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a> &copy; <a href="http://cartodb.com/attributions">CartoDB</a>'
         url="https://cartodb-basemaps-{s}.global.ssl.fastly.net/dark_all/{z}/{x}/{y}.png"
       />
-      <CircleMarker center={maroua} pathOptions={{ color: "red" }} radius={2}>
-        <Tooltip>Tooltip for CircleMarker</Tooltip>
-      </CircleMarker>
-      <CircleMarker center={garoua} pathOptions={{ color: "red" }} radius={2}>
-        <Tooltip>Tooltip for CircleMarker</Tooltip>
-      </CircleMarker>
       <CircleMarker
-        center={nagoundere}
+        center={LATLONG.maroua}
         pathOptions={{ color: "red" }}
         radius={2}
       >
         <Tooltip>Tooltip for CircleMarker</Tooltip>
       </CircleMarker>
-      <CircleMarker center={yaounde} pathOptions={{ color: "red" }} radius={2}>
-        <Tooltip>Tooltip for CircleMarker</Tooltip>
-      </CircleMarker>
-      <CircleMarker center={douala} pathOptions={{ color: "red" }} radius={2}>
-        <Tooltip>Tooltip for CircleMarker</Tooltip>
-      </CircleMarker>
       <CircleMarker
-        center={bafoussan}
+        center={LATLONG.garoua}
         pathOptions={{ color: "red" }}
         radius={2}
       >
         <Tooltip>Tooltip for CircleMarker</Tooltip>
       </CircleMarker>
-      <CircleMarker center={bamenda} pathOptions={{ color: "red" }} radius={2}>
+      <CircleMarker
+        center={LATLONG.nagoundere}
+        pathOptions={{ color: "red" }}
+        radius={2}
+      >
         <Tooltip>Tooltip for CircleMarker</Tooltip>
       </CircleMarker>
-      <CircleMarker center={buea} pathOptions={{ color: "red" }} radius={2}>
+      <CircleMarker
+        center={LATLONG.yaounde}
+        pathOptions={{ color: "red" }}
+        radius={2}
+      >
         <Tooltip>Tooltip for CircleMarker</Tooltip>
       </CircleMarker>
-      <CircleMarker center={ebolowa} pathOptions={{ color: "red" }} radius={2}>
+      <CircleMarker
+        center={LATLONG.douala}
+        pathOptions={{ color: "red" }}
+        radius={2}
+      >
         <Tooltip>Tooltip for CircleMarker</Tooltip>
       </CircleMarker>
-      <CircleMarker center={bertoua} pathOptions={{ color: "red" }} radius={2}>
+      <CircleMarker
+        center={LATLONG.bafoussan}
+        pathOptions={{ color: "red" }}
+        radius={2}
+      >
         <Tooltip>Tooltip for CircleMarker</Tooltip>
       </CircleMarker>
-      <Marker position={yaounde}>
+      <CircleMarker
+        center={LATLONG.bamenda}
+        pathOptions={{ color: "red" }}
+        radius={2}
+      >
+        <Tooltip>Tooltip for CircleMarker</Tooltip>
+      </CircleMarker>
+      <CircleMarker
+        center={LATLONG.buea}
+        pathOptions={{ color: "red" }}
+        radius={2}
+      >
+        <Tooltip>Tooltip for CircleMarker</Tooltip>
+      </CircleMarker>
+      <CircleMarker
+        center={LATLONG.ebolowa}
+        pathOptions={{ color: "red" }}
+        radius={2}
+      >
+        <Tooltip>Tooltip for CircleMarker</Tooltip>
+      </CircleMarker>
+      <CircleMarker
+        center={LATLONG.bertoua}
+        pathOptions={{ color: "red" }}
+        radius={2}
+      >
+        <Tooltip>Tooltip for CircleMarker</Tooltip>
+      </CircleMarker>
+      <Marker position={LATLONG.yaounde}>
         <Popup>
           A pretty CSS3 popup. <br /> Easily customizable.
         </Popup>
