+++ frontend/src/scripts/lat_long.ts
@@ -1,0 +1,24 @@
+
+const bamenda: any = [5.914395, 10.129316];
+const bafoussan: any = [5.468774, 10.420834];
+const bertoua: any = [4.558081, 13.662206];
+const buea: any = [4.155587, 9.232463];
+const douala: any = [4.03222, 9.706715];
+const ebolowa: any = [2.891746, 11.15648];
+const garoua: any = [9.278875, 13.394429];
+const maroua: any = [10.588261, 14.350791];
+const nagoundere: any = [7.349664, 13.577051];
+const yaounde: any = [3.826985, 11.495974];
+
+export const LATLONG = {
+    bamenda,
+    bafoussan,
+    bertoua,
+    buea,
+    douala,
+    ebolowa,
+    garoua,
+    maroua,
+    nagoundere,
+    yaounde
+}
